fix: ref without label raises NoMethodError
Bug scenario from @faresd:

> If you create a Release without a description, which is possible by
> Publishing a document and choosing "Publish at a specific date and
> time", the ref returned by Prismic's API will have a null label and the
> Prismic client will fail to parse the entire response.

Because of how the hash construction works, it needs a value for the
key. This makes it fall back to the id of the ref which should always be
present in the scenario where the label is null/nil.

fixes https://github.com/prismicio-community/ruby-kit/issues/116

More details:

- https://community.prismic.io/t/release-without-description-causes-error-in-prismic-ruby-client/7514/4